requirements.txt
mss==7.0
opencv-python==4.6.0.66
pure-python-adb==0.3.0.dev0d
Pillow==9.3.0
numpy==1.23.5